pyproject.toml
[build-system]
requires = [
    "setuptools>=64",
    "setuptools_scm>=8"
]
build-backend = "setuptools.build_meta"

[project]
name = "entity-management"
description = "Access to production entity management"
readme = { file = "README.rst", content-type = "text/x-rst" }
requires-python = ">=3.8"
license = { file = "LICENSE.txt" }
authors = [
  { name = "Blue Brain Project, EPFL" },
]
keywords = [
  "computational neuroscience",
  "simulation",
  "analysis",
  "nexus",
  "parameters",
  "BlueBrainProject",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Operating System :: POSIX",
  "Topic :: Scientific/Engineering",
  "Topic :: Utilities",
]
dependencies = [
  "attrs",
  "httpx",
  "python-dateutil",
  "rdflib",
  "pyjwt",
  "python-keycloak",
  "click",
  "jsonschema",
  "pyyaml",
  "importlib-resources; python_version <= '3.8'",
]
dynamic = ["version"]

[project.optional-dependencies]
docs = [
  "sphinx-bluebrain-theme",
]

[project.urls]
Homepage = "https://github.com/BlueBrain/entity-management"
Documentation = "https://entity-management.readthedocs.io/en/stable/"
Repository = "https://github.com/BlueBrain/entity-management"
Issues = "https://github.com/BlueBrain/entity-management/issues"

[project.scripts]
entity-management = "entity_management.cli.base:cli"

[tool.setuptools.packages.find]
include = ["entity_management", "entity_management_async", "entity_management_common"]
namespaces = false

[tool.setuptools_scm]
local_scheme = "no-local-version"

[tool.black]
line-length = 100
target-version = [
    'py38',
    'py39',
    'py310',
    'py311',
    'py312',
]
include = 'entity_management.*\.py$|tests\/.*\.py$|doc\/source\/conf\.py$|setup\.py$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-vv",
    "-ra",
    "--durations=10",
    "--durations-min=1.0",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "module"
testpaths = [
    "tests",
]
